Finish implementing user security

+++ api/src/main/java/com/dd/api/restapi/controllers/DefensivePlayerController.java
@@ -23,6 +23,8 @@ public class DefensivePlayerController {
     @Autowired
     private final TeamService teamService;
 
+    private final String exceptionMessage = "Cannot access the specified resource with the specified user id: ";
+
     @Autowired
     public DefensivePlayerController(DefensivePlayerService service, TeamService teamService, Validator validator) {
         this.service = service;
@@ -32,39 +34,60 @@ public class DefensivePlayerController {
 
     @RequestMapping("/get")
     @GetMapping
-    public DefensivePlayer get(@RequestParam Long id, @RequestParam Long userId) {
+    public DefensivePlayer get(@RequestParam Long id, @RequestParam Long userId) throws Exception {
         Objects.requireNonNull(id);
         Objects.requireNonNull(userId);
+
+        if (!this.validator.validateDefensivePlayer(userId, id)) {
+            throw new IllegalAccessException(exceptionMessage + userId);
+        }
+
         return this.service.getDefensivePlayer(id);
     }
 
     @RequestMapping("/get-by-team")
     @GetMapping
-    public List<DefensivePlayer> getByTeam(@RequestParam Long teamId, @RequestParam Long userId) {
+    public List<DefensivePlayer> getByTeam(@RequestParam Long teamId, @RequestParam Long userId) throws IllegalAccessException{
         Objects.requireNonNull(userId);
         Objects.requireNonNull(teamId);
+
+        if(!this.validator.validateTeam(userId, teamId)) {
+            throw new IllegalAccessException(exceptionMessage + userId);
+        }
+
         return this.service.getAllPlayersByTeam(teamId);
     }
 
     @RequestMapping("/create")
-    public DefensivePlayer create(@RequestBody DefensivePlayer player, @RequestParam Long userId) {
+    public DefensivePlayer create(@RequestBody DefensivePlayer player, @RequestParam Long userId) throws Exception {
         Objects.requireNonNull(player);
         Objects.requireNonNull(userId);
+
+        if (!this.validator.validateTeam(userId, player.getTeam().getId())) {
+            throw new IllegalAccessException(exceptionMessage + player.getTeam().getId());
+        }
+
         return this.service.createDefensivePlayer(player);
     }
 
     @RequestMapping("/update")
-    public DefensivePlayer update(@RequestParam Long id, @RequestBody DefensivePlayer player, @RequestParam Long userId) {
-        Objects.requireNonNull(id);
-        Objects.requireNonNull(userId);
-        Objects.requireNonNull(player);
+    public DefensivePlayer update(@RequestParam Long id, @RequestBody DefensivePlayer player, @RequestParam Long userId) throws Exception {
+        Objects.requireNonNull(id); Objects.requireNonNull(userId); Objects.requireNonNull(player);
+
+        if(!this.validator.validateDefensivePlayer(userId, id)) {
+            throw new IllegalAccessException(exceptionMessage + userId);
+        }
+
         return this.service.updateDefensivePlayer(id, player);
     }
 
     @RequestMapping("/delete")
-    public boolean delete(@RequestParam Long id, @RequestParam Long userId) {
+    public boolean delete(@RequestParam Long id, @RequestParam Long userId) throws Exception {
         Objects.requireNonNull(id);
         Objects.requireNonNull(userId);
+        if (!this.validator.validateDefensivePlayer(userId, id)) {
+            throw new IllegalAccessException(exceptionMessage + userId);
+        }
         return this.service.deletePlayer(id);
     }
 }
